Refactor and clean more tests that use tempfile, by using a safe method that always cleans up the temporary file. Also have unique prefixes to distinguish between temp files & dirs.
BUG=https://code.google.com/p/nativeclient/issues/detail?id=3631
R=bradnelson@google.com, bsy@google.com, jvoung@chromium.org

Review URL: https://codereview.chromium.org/26279009

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@12217 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 5a819eba483796989e573bc76bf6d0d391935aac

+++ tools/test_lib.py
@@ -296,7 +296,10 @@ def MakeTempFile(env, **kwargs):
   handle, path = tempfile.mkstemp()
   def Cleanup():
     try:
-      os.unlink(path)
+      # Try to remove the file but only if it exists. Some tests may clean up
+      # after themselves.
+      if os.path.exists(path):
+        os.unlink(path)
     except BaseException as e:
       sys.stderr.write('Unable to delete file %s on exit: %s\n' % (
         path, e))
